+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 - added: Added Base network support to Thorchain plugin.
+- added: Added orderUri to 0xGasless plugin EdgeTxAction metadata.
+- fixed: Fixed 0xGasless plugin orderId.
 
 ## 2.17.1 (2025-01-27)